fix(main): remove some wrapping for post.commits
I was running a match statement on the value of `post.commits` which has
the value of `Option<String>`. The correct way to use the match is
matching the value itself (just `match post.commits`) and the match arms
eliminate the need for unwrapping. Instead, I was being silly and put
the `post.commits` inside a `Some()` which gives the value
`Option<Option<String>>`. This makes the whole pattern match absurd in
the first place.

Safety has been restored.