fix(Makefile): don't export DATABASE_URL from .env
Kept running into issues where if I tried to compile the project from
the Makefile I would get a bunch of errors from SQLx but if I ran the
exact same commands in the shell it would be fine. Apparently the
problem was that SQLx expects the DATABASE_URL var to be a regular shell
variable, which it's not if it's exported by Make.

The fix is to not export the Make var and now everything works.