feat: add related summary posts and switch archive list to summaries
Added a new `SummaryPost` type which just includes some of the important
summary data of the post.

Used this new datatype to show related posts at the bottom of the page
for a post using embeddings calculated when building the database.

Changed the `/posts` page to show post summaries as well instead of the
full content.